pythonlogin/requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
python-dotenv==1.0.1
Werkzeug==3.0.2
gunicorn==23.0.0
